+++ src/tests/AutoSDK.SnapshotTests/Snapshots/elevenlabs/NewtonsoftJson/_#G.Models.RecordingResponseModel.g.verified.cs
@@ -25,7 +25,7 @@ namespace G
         /// The size of the recording in bytes.
         /// </summary>
         [global::Newtonsoft.Json.JsonProperty("size_bytes", Required = global::Newtonsoft.Json.Required.Always)]
-        public int SizeBytes { get; set; } = default!;
+        public long SizeBytes { get; set; } = default!;
 
         /// <summary>
         /// The date of the recording in Unix time.
@@ -66,7 +66,7 @@ namespace G
         public RecordingResponseModel(
             string recordingId,
             string mimeType,
-            int sizeBytes,
+            long sizeBytes,
             int uploadDateUnix,
             string transcription)
         {
