+++ src/tests/AutoSDK.SnapshotTests/Snapshots/github/SystemTextJson/_#G.Models.CodeScanningVariantAnalysisRepoTask.g.verified.cs
@@ -28,7 +28,7 @@ namespace G
         /// The size of the artifact. This is only available for successful analyses.
         /// </summary>
         [global::System.Text.Json.Serialization.JsonPropertyName("artifact_size_in_bytes")]
-        public int? ArtifactSizeInBytes { get; set; }
+        public long? ArtifactSizeInBytes { get; set; }
 
         /// <summary>
         /// The number of results in the case of a successful analysis. This is only available for successful analyses.
@@ -99,7 +99,7 @@ namespace G
         public CodeScanningVariantAnalysisRepoTask(
             global::G.SimpleRepository repository,
             global::G.CodeScanningVariantAnalysisStatus analysisStatus,
-            int? artifactSizeInBytes,
+            long? artifactSizeInBytes,
             int? resultCount,
             string? failureMessage,
             string? databaseCommitSha,
